add subscription spec with vcr

+++ app/models/subscription.rb
@@ -22,17 +22,20 @@ class Subscription < ApplicationRecord
   end
 
   def google_insert_member
+    return if channel.google_group_key.blank?
     member = Google::Apis::AdminDirectoryV1::Member.new(email: user_email)
     @@service.insert_member(channel.google_group_key, member)
   end
 
   def google_update_member
+    return if channel.google_group_key.blank?
     delivery_settings = paused? ? 'NONE' : 'ALL_MAIL'
     member = Google::Apis::AdminDirectoryV1::Member.new(delivery_settings: delivery_settings)
     @@service.update_member(channel.google_group_key, user_email, member)
   end
 
   def google_delete_member
+    return if channel.google_group_key.blank?
     @@service.delete_member(channel.google_group_key, user_email)
   end
 end
